+++ src/auth/index.js
@@ -1,0 +1,9 @@
+const { verifyToken, signToken } = require("./jwt");
+const { comparePassword, hashPassword } = require("./password");
+
+module.exports = {
+  verifyToken,
+  signToken,
+  comparePassword,
+  hashPassword,
+};