+++ src/errors/ValidationError.js
@@ -1,0 +1,9 @@
+class ValidationError extends Error {
+  constructor(message = "Invalid data provided") {
+    super(message);
+    this.name = "ValidationError";
+    this.statusCode = 400;
+  }
+}
+
+module.exports = ValidationError;