+++ src/id/index.js
@@ -1,0 +1,5 @@
+const generateUUID = require("./uuid");
+
+module.exports = {
+  generateUUID,
+};