+++ src/logging/index.js
@@ -1,0 +1,8 @@
+const { auditLog, auditLogger } = require("./audit");
+const { logger } = require("./logger");
+
+module.exports = {
+  auditLog,
+  auditLogger,
+  logger,
+};