+++ src/object/index.js
@@ -1,0 +1,5 @@
+const { deepClone } = require("./deepClone");
+
+module.exports = {
+  deepClone,
+};