+++ src/security/index.js
@@ -1,0 +1,8 @@
+const { encrypt, decrypt } = require("./encryption");
+const { sanitizeInput } = require("./sanitize");
+
+module.exports = {
+  encrypt,
+  decrypt,
+  sanitizeInput,
+};